Changed type of Product.Number to UInt16
Changed type of Product.Number to UInt16 because otherwise large product
number could not be handled

+++ LimsVisualizer/Document.cs
@@ -110,7 +110,7 @@ namespace LimsVisualizer
     public class Product
     {
         public Guid Id { get; set; }
-        public int Number { get; set; }
+        public UInt16 Number { get; set; }
         public string Name { get; set; }
         public ProductType ProductType;
     }
